refactor(stella-cli): move drive's tests to a sibling module
`drive.rs` reached 1515 lines with the probe-budget tests of #4317 on it, over
the 1500-line ceiling `make file-size` enforces. It carries no entry in
`scripts/file-size-baseline.txt`, so that is a hard failure rather than a
grandfathered one, and the remedy AGENTS.md names is the split — never a
baseline entry, which only covers files predating the guard.

A pure move: `use super::*` resolves to `drive` exactly as it did inline, and
the same five tests run. This is the shape `settings/tests.rs`,
`plugin_cmd/tests.rs` and `self_driving_cmd/learning.rs` already use.

Also deletes a content-free clause from the heartbeat test's doc comment that
`make prose` caught — "the slack is stated rather than assumed" is the
meta-writing construction CLAUDE.md names, and the sentences before it already
carry the arithmetic.

Refs #4317
